bootstrap: Add support for optional arguments in function calls
We think that we will ultimately only accept optional arguments to be
called with 'foo a b argname=c', to make it quite explicit what is
happening here.

For now, however, this is just positional.

An optional argument introduced with the following syntax:

  fun foo a:U32 ?b:?.U32 = U32
    let bval = Optional b 0
    return a + bval

Optional arguments must be nullable references. When the argument is
omitted, null is passed instead. null can of course be passed
explicitly for the same effect.

A convenience function nlang.Optional is defined:

  (fun Optional T) ptr:?.T otherwise:T = T

and is used to assign a default value when the argument is omitted or
null is passed.