bootstrap/nparser: Reorder precedence of call, cmp ops, bool ops, and tuples
As the function call syntax does not involve operators, we have to
handle precedence explicitly in the rules hierarchy. We had done so for
tuples, we needed to do the same for comparison operators and boolean
operators.

- call binds tighest (in this list)
- comparison
- boolean
- tuples

I.e one can write

  return foo a < bar b and not foo b != bar a, foo a

We don't think that's really ambiguous (especially with strong typing,
we don't have the same problems as in C where various precedence
interpretations give typable expressions).