bootstrap: Add 'extern' keyword for type, fun and method
It has no effect on fun and method at this time, but it extends the
special handling of builtin types to any user type. It means that we
will use the C struct definition directly.

A typedef must be in a module.pre.h file (for build ordering reasons).

nlang_site/posix.pre.h

	typedef struct timeval posix_timeval;
	NLANG_POINTER_ALIASES(posix_timeval);