bootstrap: Add try/catch/except support
This is not exception handling, this in fact just a clean goto.
There is no stack unwinding.

	fun error_prone = i32
	  let x = 0:u32
	    try
	      let except = nerror.OK
	      x = 1
	      let except = nerror.EINVAL
	      x = 2
	      return 0
	    catch err
	      if err != nerror.EINVAL
	        return 1
	      if x != 1
	        return 2
	  return 0

In the form
	let except = value

value must be a generalized_boolean. If operator_test__ evaluates to
true, then we goto to the catch statement and 'err' takes the
value "assigned" to 'except'. All except must unify to the same
type.

It's possible to use this for computations, and it is fast. The
only restriction is that it works only with generalized booleans,
which could be annoying at time. We could do something like:

  type except_value t = generalized_boolean
    value:t

    fun mk x:t = this
      return this{{ value=x }}

    method operator_test__ = bool
      return true

  fun trycatch_computation x:i32 = u32
    let eval32:alias = except_value u32
    try
      if x == 0
        let except = eval32.mk 1
      elif x == 1
        let except = eval32.mk 0
      elif x == 2
        let except = eval32.mk 2
      else
        return foo x
    catch eval
      return eval.value

Another approach that would have less overhead would be to
introduce a second keyword, throw (?) that doesn't use a let.

  fun trycatch_computation x:i32 = u32
    try
      if x == 0
        throw 1
      elif x == 1
        throw 0
      elif x == 2
        throw 2
      else
        return foo x
    catch eval
      return eval.value

Throw will be added in a subsequent patch.