bootstrap: [2/2] Optional arguments becomes optional named arguments
To differentiate between optional arguments and (upcoming) varargs, we
make it mandatory to pass optional arguments with their name, and
possibly out-of-order.

	fun foo a:t b:t ?option1:?@t ?option2:?@!t

	foo a b option2=@!c optiona1=@d

We choose to not go the Python route, for instance, where optional
arguments can be positional or named, and in the latter case can be
passed either by position or by name. This simplify things a lot, we
think.

The arguments names are not part of the mangled name, therefore
different compilation units can disagree on the name of an argument. We
feel that this is OK, as all arguments are positional and non-optional
at the C level: the compiler simply reorders given arguments and passes
NULL when they are missing.