Remove nlang.literal.{string,bool}, use the concrete type instead
This lets us convert a bunch of:
	'abcd':string -> 'abcd'
	true:bool -> true

This means that nlang.stringmod.string and nlang.numbers.bool are
privileged in a way other nlang.numbers.* are not
wrt nlang.literal.integer. But in the case of string and bool there are
clear defaults that do not cause any loss of information.

'a':char is special-cased in typing.unify().

We're not precluded from adding these literals back in the future, with
automatic conversion to the concrete types string and bool, in case we
want to allow custom types to use them. But it simplifies the compiler
to omit this for now.