tests_run: Remove EXCEPT traces from tests' stderr
Otherwise, function name or line changes would modify the tests.

While we could run the tests with an environment variable disabling the
print out of the trace, this would be a pain in practice, and could lead
to subtle problems (as the test runs a different code path, even though
it's supposed to be harmless). So we just filter the trace out.