bootstrap/0/lexer: Carefully handle EOF
Valid EOF are always found at a eol. re2c is designed assuming that
YYFILL(n) with n < YYMAXFILL can always be invoked successfully if the
syntax is valid. I.e. YYFILL() should in principle have a buffer not
smaller than YYMAXFILL and pad with '\0'. But we would like to use a
mmap of the file (we think), so we can't assume there is room at the end
of the buffer. Therefore, we make sure that, in effect, YYMAXFILL is 1
in the 'eol:' section of the parser.