boostrap: method wildcard mutability support
We want to avoid the typical C++ pattern where 2 accessors methods are
defined with an identical body, with a const method and a regular
method. This is error-prone and unnecessarily verbose. In N, because of
the additional mutability level, we would have 3 identical versions.

In N, we allow a special form that explicitly states "this method can be
called at all mutability levels on self". This is similar to the 2nd
order generic:

  (fun r:AnyRef) get m:(r c) = (r i32)

but on the self parameter of a method (which cannot otherwise be a
generic).

  type record =
    x:i32

  method$ get_x = @$i32
    return @$ self$x

  fun main = i32
    let r1 = record{{ x=0 }}
    let p1 = r1.get_x
    -- p1 is a const reference
    let r2 = record{{ x=1 }}
      let p2 = r2!get_x
      -- p2 is a mutable reference
      p2! = p1.

    if r1.x == r2.x
      return 0
    else
      return 1