bootstrap: do not use 'let' for type names, but 'alias'
We have been ambivalent about overloading 'let' for type names. It
presents some challenges in the ordering of passes in the compiler when
we want to get rid of forward declarations (e.g. let this = ... must be
typed before field declarations, but let MAX = this{{ x=... }} cannot be
processed at the same time because the field x is not yet typed). We
could deal with this issue by ordering typing "on demand" but that's not
the kind of complexity we wish to have at this point.

'alias' also has the advantage of standing out more in the code, which
can be helpful as type names are not otherwise easily identifiable
without some guesswork.