bootstrap: support for dyn dynamic dispatch
Add keyword 'final' to be used in intf and resolves to 'this' of the
concrete type that implements (isa) the intf.

We choose to allow the creation of dyn for all intf, even those that use
'final'. But all the members that are generic methods or functions, or
use 'final' (it is an implied generic argument after all) are simply
skipped.

We may still opt to have introduce 'static intf' that cannot be used for
dyn dispatch but that can use final. However, having to manage 2 kinds
of intf, and choose which kind is most appropriate is a pain. And there
is no really clear benefit.