bootstrap: let: group several variable definitions with 'and'
This form is useful to have a single mutating block for several
variables (without 'and', you would need several levels of nested
mutating blocks).

	let x = 0
	and y = 1
	  y = foo
	  x = bar y

Right now, there is no support for arbitrary dependency between x and y
(in ML, the 'let ... and ...' can be used for "recursive" dependencies).
x is defined strictly before y, in order of declaration. No decision has
been made on whether we would want that.