bootstrap: replace *_for_generated() family with catchup()
This rationalizes the tree rewriting process in passes, and makes it
clearer throughout the code what is being done and what assumptions are
made when rewriting a node.

It is still not entirely robust, however. It is entirely possible for
steps to mess with other surrounding steps. In essence, this comes from
the fact that while node rewriting changes the tree locally, it may
affect steps (or even passes) on nodes that are relatively far from the
rewritten node (as many steps use the node's ancestors and descendants
to act). When calling catchup(), there is no guarantee in general that
enough of the tree is re-passed to guarantee that a given node is always
rewritten the same way (steps are not commutative -- that's OK -- but
not even associative in general).

Fixing this issue may involve:
  (i) rewrite the whole tree at each pass/step, but that's expensive;
  (ii) use a COW tree structure but essentially do (i), cheaper but not
  cheap;
  (iii) adding information on each step so that pass() is aware of how
  many nearby nodes have been touched, and what kind of nodes, so that
  it can look at all the other steps and nearby nodes and decide what
  should be re-passed. This probably can be done sanely.

For reference, an example of such a subtle problem:

  FOR -- scope defines __gensym0
    LET
      DEFPATTERN
        IDENT(__gensym0)
	expr
	DEFNAME(__gensym0)
      BLOCK
        WHILE
	  ...

When introducing temporaries for the above statement, we rewrite the FOR
node into:

  LET
    DEFPATTERN
      IDENT(__gensymXX)
      expr
      DEFNAME(__gensymXX)
    BLOCK
      FOR
        LET
	  DEFPATTERN
            IDENT(__gensym0)
    	    expr
       	    DEFNAME(__gensym0)
          BLOCK
            WHILE

But in the process we could loose the scope definition for __gensym0 in
the new FOR. Recovering that information involves re-passing the tree
from the outer LET all the way to the DEFNAMEs in the inner LET. But we
only rewrote the FOR statement! In general, for other statements, such a
deep re-pass is not necessary. It can be tricky to know if enough is
re-passed in general. (iii) would work as a solution for this example,
but the logic would not be straightforward.