bootstrap: restrict generalized boolean to a test operator
We don't want to let 'and', 'or' and 'not' be redefined by generalized
booleans as we would have to break left-to-right conditional
evaluation for these operators. We may introduce operators that
evaluate both terms first (maybe 'pand', 'por', 'pnot').

Note that in C++, overloaded || and && operators are implemented with
functions and evaluate both arguments first. And we don't think that's
such a brilliant idea.

Also rename interfaces for types compatible with bool and static_string.