bootstrap: compilation support for tuples
Some cases are not yet handled, in particular: returning a tuple with
each element named, returning a tuple through reference.

We modified our approach significantly from what had been tentatively
implemented earlier on.

TYPE_TUPLE and TBI_PSEUDO_TUPLE are dropped, as we don't really want to
have a special case for a not-quite-variable-number-of-arguments-generic
generic.

Instead, we decide that it's both easier (to implement) and more
practical (to use) to define builtin generic types tuple_2 .. tuple_16
that are instantiated directly by the compiler. These definitions
automatically give the memory layout we wanted for tuples, and elements
can be accessed using the .x0 .. .x15 fields by the user. And it's also
possible to define something like:

	(type u:i_copyable v:i_copyable) pair_vector p:(tuple_2 u v) =
	  ...

(Tuples of more than 16 elements are simply not supported, and that's
OK. It would not be that hard to lift this restriction, as all tuple
types are i_any_tuple, we could allow users to define tuple_17
themselves, ...)