bootstrap: DEFNAME: explicit IDX_ to clarify special sub node
DEFNAME nodes that correspond to an 'except' contain a special subnode
that gets passed normally but is not reached through the normal flow of
statement and expressions. This subnode is interpreted as a test context
for the error, and is used to spit the error if needed by code
generators.