bootstrap: new efficient typ data structure
The unification algorithm remains the same but we now have significantly
faster implementation.

We used to link typ to typ when they unify to the one another (creating
sometimes long chains), then in a later pass "lock" the typ by rewriting
the typs with their final type.

	a -> b \
	        \
	c -> d -> e -> t

Instead, links are always of length 1 but we maintain a list of
"backlinks" that point to all the locations (of type 'struct typ *') that
need to be updated when a tentative typ is unified to a new value.

	a \
	b -> t (with backlinks to &a, &b, &c)
	c /

The global speed up so far is at least 25%.

We also compute and maintain a hash to speed up typ_equal(). We also
have a quickisa field, a hash set, to speed typ_isa(). These last two
features are not yet turned on but will provide a noticeable speedup.