bootstrap: improve cache locality in 'struct node'
Profiling with perf(1) shows that in a number of hot loops (e.g. pass())
we load 'struct node' in L2 cache in two separate loads and take a hit
waiting for each. By grouping at the front of the structure the fields
used most often, there was a (globally) noticeable improvement.