bootstrap: more generic support of bare identifiers for enums
Support in match patterns, comparisons (left and right terms), function
arguments, as before, but also in struct literals, array literals,
anywhere really.

This is now handled via an extension to the unification algorithm and is
completely transparent. This illustrates the power of this approach to
typing, as it allows delayed resolution.