bootstrap:cprinter: fix print_try: wrap catch statements in while (0) {}
It is legal C to goto inside a loop. So to prevent the no-error case in
a try from falling one of the label corresponding to a catch we wrap all
the catches in a while (0) {}: the only way to get to that code is via
an except/throw (and its generated goto).