bootstrap: fix: instantiation: do not store typ result in temp vars
With our current typ implementation, it is important to properly set
all backlinks. We were passing 'struct typ **result' to instantiating
functions, to retrieve the new instance's typ. But that can lead to all
sorts of problems if a temporary variable is passed in (including memory
corruption).

Instead, we now pass 'struct node **result', from which we can access
(*result)->typ with no problematic backlink being created.

We should reconsider our approach to prevent this kind of subtle issue.
Maybe create a 'struct typr typr' field in 'struct node' that only
exposes a 'void *' outside of types.c. If we isolate all the tricky
backlink logic in types.c some more, we should limit the damage.