bootstrap: stricter operator parentheses rules
Basically, we only allow +-*/ to be combined without parenthesis, on the
basis that most people get these rules. We also allow a sequence of a
single operator (e.g. a | b | c) to be left-associative with itself.

Anything else, in particular combining arithmetic and bitwise operators,
requires parentheses.

We still allow this: 'a == b & c', but it's not particularly dangerous
in N because of the strict typing on == (produces a bool) and &
(requires a bitwise; and bool isn't a bitwise). So really, noone should
be misinterpreting this.