bootstrap: flatten pass() into a state machine
Profiling shows pass() to be very expensive overall (30%). Indeed,
pass() involves calling itself and a large number of step function
pointers many times per nodes. A lot of the steps do nothing on most
node types, so it's a lot of function (pointer!) call overhead.

By flattening passes into state machines, we let the compiler inline the
step functions and our manual stack is likely faster than a function
callstack.

The speed up is about a little under 15%. We will next introduce a way
for a step to let the state machine now which kinds of nodes it wants to
see without having to enter the step function.