bootstrap: convert 'struct node' field 'subs' to a doubly-linked list
It is faster (we mostly save on malloc overhead) and maybe generally
more suited to our tree rewriting. But apart from allowing us to use a
simple memory pool (which in addition from performance considerations,
it guarantees all the nodes of a module can be freed at once, including
detached ones), it is not particularly immediately advantageous in terms
of how the code is written.