bootstrap: get rid of toplevel->members crutch: nest fun/method members
In the code, member functions and methods are declared outside of the
DEFTYPE. Previously, we preserved that layout in the tree but added a
reference to these "detached" members in 'struct toplevel'. This was a
bit of a nightmare, especially as for generic instances, members would
all be children of the DEFTYPE instance (as we needed somewhere to
attach them to).

Anyhoo, we now move detached members as children of their DEFTYPE as
soon as lexical scoping is done. Generic instances or not, it's all
handled the same way.