bootstrap: intf names must start with '`'
This replaces the convention of starting intf names with 'i_'. We
debated this issue at length with ourselves and came to the conclusion
that it was a welcome change. Intf do play a very large role in N, it
appears, in particular dynamic interfaces. And it seems like a good idea
to track it.

At the same time, N code is dense and presents very rich semantic
information. Are we pushing it too far?