I think we mean to use 'opaque' instead of 'opaque inline'
Inline is implicit in opaque. And while 'opaque inline' is more
explicit, and in a sense more logical, it is quite verbose for a
construct that is going to be very common in a lot of code. And it's a
construct we want to encourage. So let's keep it short.