bootstrap:types: link_to_tentative: call remove_as_user_of_generic_args()
When linking src to a tentative dst, and when src is a generic, each of
the generic arguments of src have gained a user (dst) and lost a user
(src). But in the current code src, while listed as a user in the
generic arguments, has no backlink to these elements in the users lists.

However, we did add_user() to dst. So by symmetry we need to remove src
as a user of each generic arg.

An alternative approach, more efficient, would be to use set_typ() when
adding users. Then there would be no need for either add_user() or
remove_as_user_of_generic_args(). But all the implications of that are
not yet understood.