bootstrap: more precise typ dependency and ordering
This is a complicated change.

In cprinter, we used to print out generic instances in the generic
definition's file (typically the header file), and not where the
instance is actually created. This is why the compiler did (and still
does) recompile every module, every time. For instance, whenever there
is a new vector instance (say, vector u64), lib/nlang/vectors.n.h.out
needs to be re-generated.

Now, however, for each toplevel definition, we keep track of which
generic instances it depends on quite precisely. This is done in
record_topdep() in bootstrap/passbody.c.

In cwriter, before printing out a toplevel definition, we make sure all
the generic instances it depends have been printed, and if not print
them right before their first use. Which means (vector u64) is printed
in vectors.n.h.out only if it's used by vectors.n itself. It is now
possible to avoid recompiling everything every time, although this has
not been implemented yet.

We also take advantage of the tracking to print reference instances on
demand (instead of generating them by default when printing the type
itself -- and thus missing double, triple, quadruple references...).

The tracking is still lacking, though. Indeed, it's not the same thing
if the generic instance is needed in the body of a function or in its
prototype. This is something cwriter needs to know about to handle
export/inline/opaque properly. This will come in a later change, and for
now we can only handle simple cases.

To make this change work (in the context of the upcoming environment
support), we also changed the way tentative instances are created and
tracked. This should in fact be a separate change, but the code is a tad
too intermingled to bother.

Let's explain why we had to make this change. It is possible for an
isalist typ to be of the form:

	struct a =
	  isa (`smthg this)

In such a case, when the isalist is typed in passfwd3(), the member
'this' of 'a' is not yet typed, it's given the tentative type '`any'. In
passfwd7(), 'this' is typed, and that '`any' is linked to 'a'.

So in this change we track tentative instances across passes, until
step_gather_final_instantiations() in passbody0(). They're stored in the
'struct toplevel' of the current mod->state->top_state->top.

We also make sure instances are always marked tentative directly if
requested as such (instead of making the caller itself use an additional
indirection with typ_create_tentative(), which not a proper solution
in general).

step_gather_final_instantiations() is improved (and corrected) in a
number of subtle ways as well, the details of which we leave to the
reader as an exercise.

For clarity, we also isolate the data from 'struct toplevel' that stores
the generic instances of a generic functor in a allocated-on-demand
'struct generic'.