bootstrap: mark 'inline' functions and methods in a generic type
This is not strictly correct: in fact, it is possible to have non-inline
members of a generic. In which case, they cannot be instantiated by
outside users of the module, but they can be used if the instances
already exist.

But this principle makes things difficult: first, all methods must
manually be marked inline in a type like 'vector', which is a pain;
second, the bootstrap compiler is not able to detect when such an
non-inline is not already available, and it's only caught by the C
linker, which is not pretty.

So maybe the new behavior should in fact be required by the spec.