bootstrap: optionally allow omitting const-ref operator
Because of vararg and opt arg, and because strings literals are best
passed by reference (to avoid a copy), N code tends to have a lot of
fairly pointless '@' operators. E.g.

	say @'Hello' @'World'

And the thing is, because of N's (upcoming) escape-analysis, passing a
(non-threaded) const reference to a function (that does not escape it)
is in all points equivalent to passing it by copy. So since there is no
effective difference, we now make the compiler insert the '@' operator
when it's needed in this specific case.

This is quite different from "references" in C++. In C++ the implicit
passing of addresses through references without an explicit '&' is crazy
because it is dangerous: the reference may be captured, the object
referenced may be modified, etc -- even though the code reads like a
pass-by-value, and it's natural to assume that's what it means. So in
C++, when reading code that is known to sometimes use references, one
must be very paranoid and check function definitions to make sure
nothing stupid is being done.

In N we only allow implicit '@' operators in the case where the code has
exactly the same effective semantics as pass-by-value. If we cannot
guarantee this property, then we'll have to undo this change. (But we
think the full language will be capable of offering such a guarantee.)

We may later decide to disallow an explicit '@' in this specific case.
Indeed, we may want to avoid the practice of inserting "unnecessary" '@'
to be a matter of style. These 2 lines do the exact same thing, and it's
ugly that they are both legal:
	say @'Test'
	say 'Test'
We do prefer the second line, so we may want to enforce this preference
in the compiler.