bootstrap: syntax: allow same-line 'and' and 'such' for let
The extra line that was always necessary before 'and' and 'such' was a
bit of a pain at times. In general, we don't want to make single-line
expressions a common thing, but in this case, especially for 'such', it
is visually helpful.

This does introduce some matter-of-style aspect to the syntax.

O, the horror of relying on other people's taste.