bootstrap: [1/3] constraint support: ssa-ish form
The support for semantic constraints is coming in in an haphazard manner.
A lot of code is being changed all over the code base as we figure out
the best approach to do this.

The trickiest issue with semantics constraints is that we now very much
need to care about control flow.

Our node tree representation was very close to the code structure
as-written, but that didn't make branches explicit at all (esp. throw,
break, ...).

In this series of changes we progressively migrating to a SSA-ish form
where the control graph is apparent.

We also introduce PHI nodes that help us keep track of the use of
local variables throughout a function, and especially in branches.
Semantic constraints apply to the values taken by a variable, unlike
types that apply to variables themselves.