bootstrap/table: Use uppercase consistently in macro names; pass hashf/cmpf to macros
hashf and cmpf where stored as function pointers in the hash tables, but
this really makes no sense. There is no need to have different hash and
comparison functions for a given hash table key_type,value_type
combination.

This should be marginally faster as now the hashf and cmpf calls can be
inlined.