bootstrap: [2/3] constraints support (LIR form)
The parser produces HIR (High-level intermediate representation), which
closely follows the code as-written. It's a bit like an AST.

The LIR pass, together with the SSA pass, converts the HIR to a
simpler representation. This allows the rest of the compiler that
only ever interacts with LIR to make stronger assumptions about the
node structure, and simplifies the compiler.

The control flow is easier to analyze in this representation.

The constraint support in this change is sufficient to support the
builtin 'nonnull' and 'init' constraints in all the current tests (with
the exception of t03/constraints.n, which will come next).