bootstrap: [3/3] constraints support (unions)
The support is still far from complete, of course, but we can now run
all the tests.

This commit adds support for tracking the possible tags of a union
value.

We have now made the clear decision that the constraint system is a
pure labeling mechanism, i.e. it is possible to compile a program whose
constraints have been elided. This is quite important as we may choose
to support constraints that are undecidable in general.

As a consequence, we cannot use the following forms
	let u = {x=1}:un::|A
	u.x = 1
because the chosen tag must be explicit somewhere else than in the
constraints:
	let u = {x=1}:u.A
	u!A.x = 1