bootstrap: expand automagic reference and dereference
We recently started allowing the implicit taking of unqualified const
references (which will mean non-capturing in the future), and in fact
made it a requirement. It allows the naked:
	say 'test'

It only makes sense to extend this to other syntax forms, since they're
supposed to be syntactic sugar for function calls (and actually are
for types that implement `arithmetic, say).
	let a = 'test'
	and b = 'test'

	let x = a == b
	assert x
	assert <- @a == b
	assert <- a == @b
	assert <- @a == @b

	let n = @1
	let m = @2
	assert <- n + m == 3:i32
	assert <- @1 + @2  == 3:i32
	assert <- @1 + @2  == @3:i32
	assert <- 1 + @2  == @3:i32
	assert <- 1 + @2  == 3:i32

Since we have distinct reference comparison operators === and !==, and
that there is no pointer arithmetic in N, this is not ambiguous.

But it could be argued that it is confusing.