bootstrap: unify: change handling of reference compatibility
There are many cases where N allows implicit conversion of references.
For instance:

  fun optional p:?@i32 default:i32 = i32
    if p === null
      return default
    else
      return p.

  let p = @1
  let x = optional p 0

The call to 'optional' requires a conversion from @i32 to ?@i32.

Until now, this was handled in a manner that was transparent to most of
the unification algorithm. For the sake of unification, the type of the
first function argument would be changed to (__ref_compatible ?@i32).
When presented with unify(__ref_compatible ?@i32, @i32), the code would
enter a special case and:
 - check that a reference with functor ?@ accepts an argument with
   functor @ by applying the reference compatibility rules,
 - directly unify(i32, i32).

This was in fact made more complicated by the handling of tentative
types and interfaces.

However, this would break with generic functions that accept more
complex arguments of types like @@i32. The artificial __ref_compatible
could end up leaking in the generic arguments of the function instance
as if it were a real type.

So, instead, we adopt a different approach with a second entry point to
the unification: unify_refcompat(). This functions treats its first
argument differently, as the target of a reference assignment.

We now have to carry a new 'flags' argument in parts of the unification,
to encode which one of the type arguments tolerates reference
conversions, if any. (Indeed, references can only be converted in
specific cases.)

do_unify() remains otherwise symmetric in 'a' and 'b'.