trun: exiting with 0 while producing the expected non-empty stderr is a pass
With this change, we avoid a spurious
	FAILED t00/fixme05.n: did not fail (exit 0)
for tests that produce a stderr but do not exit with non-zero (e.g.
by writing non-fatal warnings to stderr).

We take the risk, however, that we will fail to detect tests that stop
returning non-zero exit code while producing the same stderr output.

We should really record the exit code too.