bootstrap: infer trivial interfaces (ctor, dtor, copy, equality, order) and `return_by_copy
Infer these interfaces by default (unless the types of the field
disallow it, or a method definition is given).

That is, if all fields are `trivial_copy, then the struct is inferred to
be `trivial_copy.

In the near future, we will instead *assume* these interfaces, and
allow users to counter this assumption by implementing a
`nontrivial_ctor or `nontrivial_copy interface (say).

For this approach to be sane, we will need to infer interfaces like
`trivial_copy_but_owned (or maybe named `capturing_trivial_copy).