bootstrap: support isalist literals such as {`trivial_copy `trivial_ctor}
	alias `both = {`trivial_copy `trivial_ctor}
	intf `a =
	  isa `trivial_copy `trivial_ctor

`both and `a are different: `both is exactly the same as the combination
of the two interfaces `trivial_copy and `trivial_ctor, no more, no less.
I.e.
	a isa `both <=> a isa `trivial_copy and a isa `trivial_ctor

`a is however something more. While to implement `a you have to
implement `trivial_copy and `trivial_ctor, a new interface is being
defined, which can involve additional constraints. Ideally, these
additional requirements are somehow captured in code (semantic
constraints, examples), but they may only be in comments or in an
external document. I.e.

	a isa `a => a isa `trivial_copy and a isa `trivial_ctor