bootstrap: shallow accessor methods
We had added the 'method@$' wildcard feature ostensibly to support
accessor methods, allowing the use of the same method name with
an implicit generic on the kind of reference for self.

This first approach was a little too simplistic however, as it failed to
consider the question of deep vs. shallow constness.

There are multiple combinations of mutability that make sense for an
accessor method. It depends on the nature of 'self' ownernship of the
data being returned.

Deep constness: (like a struct field)
	t method@ get = @i32
	t method@! get = @!i32
	t method@# get = @#i32
(this is the current behavior of method@$).

Shallow constness: (like a container)
	t method@ get = @i32
	t method@! get = @#i32

Very shallow constness (just a regular method; or like a struct pointer
field):
	t method@ get = @#i32

In this change, we introduce a new kind of wildcard method:

	t shallow method@$ get = @$i32

It corresponds to
	t method@ get = @i32
	t method@! get = @#i32
	( t method@# get = @#i32 )

In a shallow wildcard method, $ acts differently on self than on other
expressions. That is, it is "more const" on self.

Containers will typically use shallow wildcard methods. While methods
mimicking struct fields will use regular wildcard methods.