[PATCH 5/15] improve instantiation completion logic
Record all instances, even tentative, in the current top node generic
instances list. Then, when completing instantiations, skip instances in
that list that are *still* tentative by that point. This is necessary as
instances may become final quite a bit later than when the instances are
first created; and then we do need to complete them.