[PATCH 7/15] NODE__DETACHED flag for nodes that are not quite "in the tree"
Generic instances are not full members of the tree: they have a parent
(the parent of the generic node), but no sibling. The reason is that
while it is helpful to give a perfectly reasonable parent to these
nodes, we don't want the instances to appear in the regular passes
through the tree: instances that remain tentative should be skipped.

We go through instances using a special-purpose logic found in
step_complete_instantiation(). Or in cprinter, using the topdep
mechanism.

It may be a good idea to abandon this strategy down the road, and simply
learn to skip tentative instances (or for cprinter, learn to skip all
generic instances, only getting to them through topdep), but that would
be a significant rework.

This patch merely keeps track of which nodes have this property, so we
can strengthen the integrity checks in node_invariant() for regular,
non-detached nodes.