[PATCH 8/15] fix operator precedence handling
An error in the encoding of the operators array in lexer.h was hiding a
bug in the handling of left associative operators. Also, when the
precedence argument of p_*() function was changed to a enum token_type,
a number of mistakes were made -- which were no caught by the
non-existent type checking on enums.